+++ src/cpu/CMakeLists.txt
@@ -44,16 +44,17 @@ add_subdirectory(hamaji)
 add_subdirectory(mayah)
 add_subdirectory(munetoshi)
 add_subdirectory(peria)
-add_subdirectory(tzik)
-add_subdirectory(yamaguchi)
-add_subdirectory(test_lockit)
-add_subdirectory(yuricat)
 add_subdirectory(ryumas)
 add_subdirectory(takapt)
+add_subdirectory(test_lockit)
+add_subdirectory(tzik)
+add_subdirectory(yamaguchi)
+add_subdirectory(yuricat)
 
 add_subdirectory(sample)
-add_subdirectory(sample_suicide)
+add_subdirectory(sample_beam)
 add_subdirectory(sample_rensa)
+add_subdirectory(sample_suicide)
 
 add_subdirectory(control_tester)
