+++ src/cpu/peria/CMakeLists.txt
@@ -25,8 +25,6 @@ peria_add_executable(peria main.cc)
 peria_add_executable(tokopuyo tokopuyo.cc)
 
 cpu_add_runner(run.sh)
-cpu_add_runner(quick_full.sh)
-cpu_add_runner(quick_2dub.sh)
 
 cpu_add_runner(book.txt)
 
@@ -36,7 +34,3 @@ include_directories(${gtest_SOURCE_DIR}/include
 peria_add_test(pattern_test)
 #peria_add_test(field_test)
 #peria_add_test(player_test)
-
-# ---------------------
-cpu_add_executable(beam_search_ai beam_search_ai.cc)
-cpu_target_link_common_libraries(beam_search_ai)
