+++ src/cpu/sample_beam/CMakeLists.txt
@@ -1,0 +1,16 @@
+cmake_minimum_required(VERSION 2.8)
+
+cpu_setup("sample_beam")
+
+puyoai_add_cxx_flags("-Wall -Wextra")
+puyoai_add_cxx_flags("-Wno-error=unused-parameter -Wno-error=unused-variable")
+
+function(sample_beam_add_executable exe)
+  cpu_add_executable(${exe} ${ARGN})
+  cpu_target_link_common_libraries(${exe})
+endfunction()
+
+sample_beam_add_executable(beam_search_ai beam_search_ai.cc main.cc)
+
+cpu_add_runner(run_full.sh)
+cpu_add_runner(run_2dub.sh)
