+++ src/cpu/sample_beam/run_2dub.sh
