+++ src/cpu/sample_beam/run_full.sh
